[SCRUM-10] Rename : PageResponse global로 이동
